Fix: Final verification and fixes for company registration flow (CountryRepository ID fix and documentation updates)

+++ Ginuma_new_updates/ginum-backend-main/src/main/java/com/example/GinumApps/repository/CountryRepository.java
@@ -7,8 +7,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-
 @Repository
-public interface CountryRepository extends JpaRepository<Country, Long> {
+public interface CountryRepository extends JpaRepository<Country, Integer> {
     Optional<Country> findByName(@NotNull String name);
 }
